V0.15.1 (#557)
* bump version and move installer to 0.15.1

* brings sema json analysis up to modern zig

* full export of erlnif functions

* changes callconv symbol

* most tests passing, except for commented-out tests

* undisable disabled tests

* un-skip key tests

* transition to segregating code via compilation model and using Sema protocol

* switches over to fully templated semantic analysis

* fixes templating error in sema_c

* re-enable advanced allocation strategies

* re-enable disabled guides

* restores allocators and c guides

* routes sema through the new build.zig template

* purge old build.zig and semantic analysis systems

* removes compiler errors

* refactor to use 'module' as the keyword for extra modules

* fixes module specification errors

* rename from modules to extra_modules

* activates project dependencies

* completes build zig override and options normalization

* stashes semantic analysis pass in its own linksection in the shared object

* adds in the test for precompilation

* splits out sema building from sema execution

* actual test conditions on precompiled_test

* obtains precompiled data from stashed semantic analysis json

* adds basic precompiled module test

* enable web retreival of precompiled assets

* adds guards on FORCE_RECOMPILE and FORCE_RELOAD

* adds web-based precompilation

* precompiled documentation in the guides

* updates optimize and error_tracing options

* adds test on recursive resources

* adds get_list_cell

* documentation updates

* fix the zig version command for linux

* workflow updates macos/windows

* compat for versions without :json

* make sure sema json is a single binary

* remove mcp tools

* compat for Elixir version < 1.16

* force trace for debug

* disable build_zig_override_test

* format and temporarily remove another test

* vibecoded macos sema_json

* resolves memory-mapping segfault in linux

* don't forget to clear local precompiled

* fixes embedded sema test strategy

* remove trace and bump freebsd llvm to 20

* normalize use of json across elixir versions

* disables testing multiplatform if the erts version is too low

* moves error return test

* MacOS errors resolved

* common objcopy for windows, freebsd

* add windows include dir

* make sure we're using dll for windows

* switches llvm to 20, freebsd

* updates with win64 cross-compilation capability

* disable build_files_test and reenable compile_error

* fixes freebsd to use stable elixir

* freebsd objcopy invocation

* freebsd objcopy invocation

* error message normalization

* path normalization for windows

* repeat json

* rror message normalization

* fix precompiled_sema command

* remove skip_windows

* fix get_list_cell

* disable error return traces in windows (for threaded)

* fixes? windows dependency path issue

* windows: fixes staging dir error

* fixes erlang error test

* prevent stack tracing bug in GPA for windows

* validate directory existence first

* repairs staging dir test

* renames extramodule build module

* bump installer to 0.15.2

* select 0.15.2 for zig_get

* general dependency lock update

* 0.19.1 update and add precompiled assets

* fix precompiled tests with new uploads for 0.15.2

* fixes macos tmp drive issue by forcing symlink resolution

+++ test/test_helper.exs
@@ -11,12 +11,12 @@ Logger.configure(level: log_level)
 
 custom_directory = "test/.custom_location"
 
+if :os.type() == {:win32, :nt} do
+  File.mkdir_p!(System.tmp_dir!())
+end
+
 if File.dir?(custom_directory), do: File.rm_rf!("test/.custom_location")
 File.mkdir_p!("test/.custom_location")
-
-defmodule MyApp do
-  def env, do: :test
-end
 
 ZiglerTest.Compiler.init()
 
@@ -28,4 +28,7 @@ ZiglerTest.MakeReadme.go()
 
 ZiglerTest.MakeZig.go()
 
-ExUnit.start()
+exclude_more = List.wrap(if Version.match?(System.version(), "~> 1.19"), do: :on_upgrade)
+exclude = ExUnit.configuration()[:exclude] ++ exclude_more
+
+ExUnit.start(exclude: exclude)
